Added scripts debug mode

+++ scripts/build-bestiary-letter.sh
@@ -4,7 +4,12 @@
 
 if [ -z "$1" ]; then
 	echo "Usage: build-bestiary-letter.sh path/to/cairn/dir"
+	echo "Any second parameter keeps the temporary directory"
 	exit 1
+fi
+
+if [ -z "$2" ]; then
+	echo "DEBUG MODE"
 fi
 
 # Directories set up
@@ -51,4 +56,9 @@ sed -i '$a \\\end{document}' $tmpdir/cairn-bestiary.tex
 pdflatex -interaction=nonstopmode -output-directory=$tmpdir $tmpdir/cairn-bestiary.tex 
 pdflatex -interaction=nonstopmode -output-directory=$tmpdir $tmpdir/cairn-bestiary.tex
 mv $tmpdir/cairn-bestiary.pdf "$destdir/cairn-bestiary-letter.pdf"
-rm -rf $tmpdir
+
+if ! [ -z "$2" ]; then
+	rm -rf $tmpdir
+else
+	echo "Build files kept in $tmpdir"
+fi
