+++ scripts/build-bestiary-lulu.sh
@@ -3,10 +3,14 @@
 # Parameters validation
 
 if [ -z "$1" ]; then
-	echo "Usage: build-bestiary-lulu.sh path/to/cairn/dir"
+	echo "Usage: build-bestiary-lulu.sh path/to/cairn/dir [debug]"
+	echo "Any second parameter keeps the temporary directory"
 	exit 1
 fi
 
+if [ -z "$2" ]; then
+	echo "DEBUG MODE"
+fi
 # Directories set up
 
 BASE_DIR=$1
@@ -51,4 +55,9 @@ sed -i '$a \\\end{document}' $tmpdir/cairn-bestiary.tex
 pdflatex -interaction=nonstopmode -output-directory=$tmpdir $tmpdir/cairn-bestiary.tex 
 pdflatex -interaction=nonstopmode -output-directory=$tmpdir $tmpdir/cairn-bestiary.tex 
 mv $tmpdir/cairn-bestiary.pdf "$destdir/cairn-bestiary-lulu-interior-$currentdate.pdf"
-rm -rf $tmpdir
+
+if ! [ -z "$2" ]; then
+	rm -rf $tmpdir
+else
+	echo "Build files kept in $tmpdir"
+fi
